% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi}
\alias{psi}
\title{Calculates the wave function of the hydrogen atom}
\usage{
psi(n, l, m, r, theta, phi)
}
\arguments{
\item{n}{The principle quantum number, n}

\item{l}{The azimuthal quantum number, l}

\item{m}{The magnetic quantum number, m}

\item{r}{The radius of the spherical coordinates}

\item{theta}{The angle θ of the spherical coordinates}

\item{phi}{The angle φ of the spherical coordinates}
}
\value{
The value of ψ calculated with the hydrogen wavefunction equation
}
\description{
Calculates the wave function of the hydrogen atom using the
Boost implementation of the spherical harmonics. Note that this function
returns psi and not the absolute value of psi squared.
}
\examples{
psi(2, 1, 0, 0.5, 2, 5)
psi(4, 2, -1, 0.7, pi/2, pi)
}
