% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavefunction.R
\docType{package}
\name{wavefunction}
\alias{wavefunction}
\title{Calculate the Hydrogen Wave Function}
\description{
Calculate the wave function of the hydrogen atom using the
Boost implementation of the spherical harmonics.
}
\details{
wavefunction

Calculate the wave function of the hydrogen atom using the Boost
implementation of the spherical harmonics.

The \emph{wavefunction} package calculates the wave function of the
hydrogen atom using the Boost implementation of the spherical harmonics. Note
that the \link[wavefunction]{psi} function returns the value of psi
and not that of the absolute value of psi squared.
}
\seealso{
\link[wavefunction]{psi}
}
\author{
Hayden Leatherwood \email{hmleatherwood@gmail.com} - 2022
}
